requirements.txt
sentence-transformers
scikit-learn
pandas
numpy
streamlit
sqlalchemy
psycopg2-binary
python-docx
pypdf
